requirements.txt
python-pptx
python-dotenv
huggingface_hub
requests